Fix integration tests for release.
