+++ test/integration/release/fails-if-master-is-behind/first.txt
@@ -1,0 +1,1 @@
+first file
