+++ test/integration/release/fails-on-pending-changes/expected.js
@@ -4,6 +4,6 @@ const exitCode = 1;
 
 const stdout = ``;
 
-const stderr = '';
+const stderr = '✗ You have pending changes, commit your changes first.';
 
 module.exports = { exitCode, stdout, stderr };
